+++ packages/core/src/content-stream/operators/graphics-state/line-join-handler.basic.test.ts
@@ -3,7 +3,7 @@ import type { PdfObject } from "../../../pdf/types/pdf-types/index";
 import { GraphicsStateStack, LineJoin } from "../../graphics-state/index";
 import { OperandStack } from "../../operand-stack/index";
 import type { OperatorHandlerContext } from "../../operator-registry/index";
-import { lineJoinHandler } from "./line-join";
+import { lineJoinHandler } from "./line-join-handler";
 
 const buildContext = (operands: PdfObject[]): OperatorHandlerContext => {
   const operandStack = OperandStack.create();
