+++ packages/core/src/content-stream/operators/graphics-state/miter-limit-handler.basic.test.ts
@@ -3,7 +3,7 @@ import type { PdfObject } from "../../../pdf/types/pdf-types/index";
 import { GraphicsStateStack } from "../../graphics-state/index";
 import { OperandStack } from "../../operand-stack/index";
 import type { OperatorHandlerContext } from "../../operator-registry/index";
-import { miterLimitHandler } from "./miter-limit";
+import { miterLimitHandler } from "./miter-limit-handler";
 
 const buildContext = (operands: PdfObject[]): OperatorHandlerContext => {
   const operandStack = OperandStack.create();
